Use IO::Select and read pipe as soon as possible - no temp files
This basically reverts d37ec03a8b7b7e0a831d4cb4c1fe31632483990b and
its temp files.

So compared to V_0.05, instead of waiting for the process to finish
before reading its pipe (which caused the deadlock), we now read the
pipe and don't expect the process to end before we've read it. This is
cleaner than temp files

Thanks to Marko Nordberg for his efforts on this